# src/karmen_backend/openapi/swagger.yaml
openapi: 3.0.0
info:
  version: '0.9.0'
  title: 'Karmen API'
  description: 'API of https://github.com/fragaria/karmen'
...
paths:
...
  '/organizations/{org_uuid}/users':
    post:
      summary: Add user to organization by his email
      description: If user exists, he is added to the org. Otherwise, new account is created and invitation is sent to him
      operationId: server.routes.users.add_user_to_org
      tags:
        - users
      parameters:
        - $ref: '#/components/parameters/organization_uuid_param'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              properties:
                email:
                  type: string
                  format: email
                role:
                  type: string
                  enum: [user, admin]
      responses:
        '200':
          description: ''
        '400':
          description: Unable to create or add user
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/api_err_response'
              examples:
                Missing email:
                  value:
                    message: Missing email
                Bad role:
                  value:
                    message: Bad role
                Invalid email:
                  value:
                    message: Invalid email
    get:
      summary: 'Get list of users in an organization'
      operationId: server.routes.users.list_users
      tags:
        - users
      parameters:
        - $ref: '#/components/parameters/organization_uuid_param'
      responses:
        '200':
          description: 'List of users'
  '/organizations/{org_uuid}/users/{user_uuid}':
    patch:
      summary: Change user's role in an organization
      operationId: server.routes.users.update_user
      tags:
        - users
      parameters:
        - $ref: '#/components/parameters/organization_uuid_param'
        - $ref: '#/components/parameters/user_uuid_param'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              properties:
                role:
                  type: string
                  enum: [user, admin]
      responses:
        '200':
          description: 'Role changed'
        '400':
          description: 'Bad role'
    delete:
      summary: 'Remove user from an organization'
      operationId: server.routes.users.delete_user
      tags:
        - users
      parameters:
        - $ref: '#/components/parameters/organization_uuid_param'
        - $ref: '#/components/parameters/user_uuid_param'
      responses:
        '204':
          description: 'User removed'
        '409':
          description: Cannot remove yourself
...
components:
  parameters:
...
    organization_uuid_param:
      name: org_uuid
      in: path
      required: true
      schema:
        $ref: '#/components/schemas/uuid'
...
    user_uuid_param:
      name: user_uuid
      in: path
      required: true
      schema:
        $ref: '#/components/schemas/uuid'
...
  schemas:
    uuid:
      type: string
      format: uuid
      pattern: '[0-9a-fA-F]{8}\-[0-9a-fA-F]{4}\-[0-9a-fA-F]{4}\-[0-9a-fA-F]{4}\-[0-9a-fA-F]{12}'
...
    api_err_response:
      type: object
      properties:
        message:
          type: string
...